// build-a-bot/src/store/mutation-types.ts
enum MutationTypes {
  ADD_ROBOT_TO_CART = 'ADD_ROBOT_TO_CART',
  UPDATE_HEAD_PARTS = 'UPDATE_HEAD_PARTS',
  UPDATE_ARM_PARTS = 'UPDATE_ARM_PARTS',
  UPDATE_TORSO_PARTS = 'UPDATE_TORSO_PARTS',
  UPDATE_BASE_PARTS = 'UPDATE_BASE_PARTS',

  UPDATE_CURRENT_USER = 'UPDATE_CURRENT_USER',
}

export default MutationTypes;